+++ index.js
@@ -2,6 +2,7 @@ require('dotenv').config()
 const schedule = require('node-schedule');
 const { App } = require('@slack/bolt');
 const logic = require('./logic');
+const home = require('./home')
 
 const app = new App({
   token: process.env.SLACK_BOT_TOKEN,
@@ -21,137 +22,25 @@ app.event('reaction_added', async ({ event, client }) => {
 });
 
 app.event('app_home_opened', async ({ event, client }) => {
-  updateHome(client, event.user);
+  home.update(client, event.user);
 });
 
 app.action(`toimistolla_click`, async ({ body, ack, client}) => {
   logic.setInOffice(body.user.id, body.actions[0].value)
-  updateHome(client, body.user.id);
+  home.update(client, body.user.id);
   await ack();
 });
 
 app.action(`etana_click`, async ({ body, ack, client}) => {
   logic.setAsRemote(body.user.id, body.actions[0].value)
-  updateHome(client, body.user.id);
+  home.update(client, body.user.id);
   await ack();
 });
 
 app.action(`update_click`, async ({ body, ack, client}) => {
-  updateHome(client, body.user.id);
+  home.update(client, body.user.id);
   await ack();
 });
-
-const updateHome = async (client, userId) => {
-  const date = new Date()
-  const days = logic.generateWeek(date)
-  let dayBlocks = []
-
-  dayBlocks = dayBlocks.concat(
-    {
-      "type": "section",
-      "text": {
-          "type": "plain_text",
-          "text": `Tiedot päivitetty ${date.toLocaleString()}`
-        }
-    },
-    {
-      "type": "actions",
-      "elements": [
-          {
-            "type": "button",
-            "text": {
-              "type": "plain_text",
-              "emoji": true,
-              "text": "Päivitä"
-            },
-            "value": "updated",
-            "action_id": 'update_click'              
-          }
-      ]
-    },
-    {
-      "type": "divider"
-    }
-  )
-
-  days.forEach(d => {
-    dayBlocks = dayBlocks.concat(
-      {
-        "type": "header",
-        "text": {
-            "type": "plain_text",
-            "text": d
-          }
-      }
-    )
-    
-    const enrollments = logic.getEnrollmentsFor(d)
-    let usersString = enrollments.length === 0 ? "Kukaan ei ole ilmoittautunut toimistolle!" : "Toimistolla aikoo olla:\n"
-    enrollments.forEach((user) => {
-      usersString += `<@${user}>\n`
-    })
-
-    dayBlocks = dayBlocks.concat(
-      {
-        "type": "section",
-        "text": {
-            "type": "mrkdwn",
-            "text": usersString
-          }
-      }
-    )
-    
-    dayBlocks = dayBlocks.concat(
-      {
-        "type": "section",
-        "text": {
-            "type": "mrkdwn",
-            "text": "Oma ilmoittautumiseni:"
-          }
-      },
-      {
-        "type": "actions",
-        "elements": [
-            {
-              "type": "button",
-              "text": {
-                "type": "plain_text",
-                "emoji": true,
-                "text": "Toimistolla"
-              },
-              "style": `${logic.userInOffice(userId, d) ? "danger" : "primary"}`,
-              "value": d,
-              "action_id": 'toimistolla_click'              
-            },
-            {
-              "type": "button",
-              "text": {
-                "type": "plain_text",
-                "emoji": true,
-                "text": "Etänä"
-              },
-              "style": `${logic.userIsRemote(userId, d) ? "danger" : "primary"}`,
-              "value": d,
-              "action_id": 'etana_click'
-          }
-        ]
-      },
-      {
-        "type": "divider"
-      }
-    )
-  })
-
-  const blocks = dayBlocks
-
-  client.views.publish({
-    user_id: userId,
-    view: {
-       type:"home",
-       blocks: JSON.stringify(blocks)
-    }
-  })
-}
 
 (async () => {
   await app.start(process.env.PORT || 3000);
